pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["pavel shevchenko <litvin.para@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-debug-toolbar = "^4.3.0"
jinja2 = "^3.1.3"
psycopg2 = "^2.9.9"
faker = "^24.4.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
